feat(firewall): dynamic port ranges, first-class opaque deny, deny observability
Make EgressRule.port a dynamic single-or-range string and add first-class
opaque deny enforcement with deny-always-wins conflict resolution.

Port model (zero drift: config schema = on-disk rules = adminv1 proto):
- EgressRule.Port is now a string ("443" or "9000-9100"); PortRange deleted.
  adminv1 port + FirewallRemoveRuleRequest.port uint32 -> string.
- config.ParsePortSpec validates bounds (1..65535) + ordering; invalid specs
  are dropped fail-closed in NormalizeAndDedup.

Conflict resolution (resolveOpaquePortConflicts):
- same-action overlapping spans merge (a range + an in-range single no longer
  emit a duplicate dedicated listener -> the bug that broke on first use).
- DENY ALWAYS WINS: deny spans carve out of allow spans, both directions,
  range-vs-range included; every denied port becomes an explicit deny rule.
- an all-single allow/deny clash on one port has no range to carve and FAILS
  LOUD at generation (checkOpaquePortActionConflicts) instead of collapsing.

First-class opaque deny chains:
- tcp/ssh/udp deny rules build a dedicated / prefix_ranges listener ->
  deny_cluster blackhole + action=denied access log, plus an eBPF route to the
  deny listener (active reset + observable, not silent default-deny).

Span-aware collision guard:
- checkProtoCollisions compares port SPANS per host across canonical protos
  (counting opaque-deny stacks), so tcp:4242-5000 + ssh:4500 fails cleanly.

Deny observability:
- the catch-all deny floor (denyDefaultFilterChain) now emits action=denied, so
  the most common deny (anything not explicitly allowed) is captured in the
  event stream instead of being silently reset.

Docs: deny semantics documented in .claude/rules/envoy.md; ENVOY_TARGET.md
(refactor-tracking scaffolding) deleted + references scrubbed; CLI reference and
configuration.mdx regenerated.

Build: bump pinned Go 1.25.10 -> 1.25.11 (clears stdlib advisories GO-2026-5039
net/textproto + GO-2026-5037 crypto/x509, both fixed in 1.25.11; SCA gate was
blocking the branch) and fold in pending dependabot bumps -- otel 1.43->1.44
(+sdk/metric/trace), otel/log + sdk/log + otlplog/otlploggrpc 0.19->0.20,
genproto/googleapis/rpc, modernc.org/sqlite 1.50.1->1.51.0, go-colorable
0.1.14->0.1.15; Dockerfile golang base -> 1.25.11-alpine@sha256:cd2fb35.

Verified live (in-agent UAT): range -> two STATIC listeners with correct
route_map projection; out-of-range port denied at the floor with an
action=denied event captured in clawker-envoy. Follow-up: GH #329.

+++ internal/controlplane/firewall/handler.go
@@ -697,7 +697,7 @@ func (h *Handler) FirewallRemoveRule(ctx context.Context, req *adminv1.FirewallR
 	rule := config.EgressRule{
 		Dst:   req.GetDst(),
 		Proto: req.GetProto(),
-		Port:  int(req.GetPort()),
+		Port:  req.GetPort(),
 	}
 	pathMode := req.GetPath() != ""
 	var matched bool
@@ -1306,7 +1306,7 @@ func ProtoRulesToConfig(in []*adminv1.EgressRule) []config.EgressRule {
 		out = append(out, config.EgressRule{
 			Dst:         r.GetDst(),
 			Proto:       r.GetProto(),
-			Port:        int(r.GetPort()),
+			Port:        r.GetPort(),
 			Action:      r.GetAction(),
 			PathRules:   paths,
 			PathDefault: r.GetPathDefault(),
@@ -1328,7 +1328,7 @@ func ConfigRulesToProto(in []config.EgressRule) []*adminv1.EgressRule {
 		out = append(out, &adminv1.EgressRule{
 			Dst:         r.Dst,
 			Proto:       r.Proto,
-			Port:        uint32(r.Port),
+			Port:        r.Port,
 			Action:      r.Action,
 			PathRules:   paths,
 			PathDefault: r.PathDefault,
